requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
redis==4.6.0
django-redis==5.2.0
django-ckeditor==6.7.0
django-resized==1.0.6
django-compressor==4.4
django-cleanup==8.0.0
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
django-ckeditor-5==0.2.10  # Alternative if we want CKEditor 5